types: use fixed-width int for uint32/int32 on LP64
On 64-bit platforms 'unsigned long'/'long' are 64-bit, so uint32/int32
were oversized. Map them to 'unsigned int'/'int', and make int8 an
explicit 'signed char' (plain char signedness is implementation-defined).

+++ include/types.h
@@ -32,11 +32,11 @@
 
 typedef unsigned char boolean;
 typedef unsigned char uint8;
-typedef char int8;
+typedef signed char int8;
 typedef unsigned short uint16;
 typedef short int16;
-typedef unsigned long uint32;
-typedef long int32;
+typedef unsigned int uint32;
+typedef int int32;
 typedef unsigned long long uint64;
 typedef long long int64;
